InstagramFurniture/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astunparse==1.6.3
autopep8==1.6.0
beautifulsoup4==4.9.3
black==22.3.0
bs4==0.0.1
cachetools==4.2.2
click==8.1.2
distlib==0.3.2
Django==2.2
django-debug-toolbar==3.2.4
django-silk==4.2.0
djangorestframework==3.13.1
filelock==3.0.12
flake8==4.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gprof2dot==2021.2.21
grpcio==1.34.1
h5py==3.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
llvmlite==0.36.0
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numba==0.53.1
numpy==1.19.5
opencv-python==4.5.2.54
opt-einsum==3.3.0
orjson==3.6.5
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.17.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==1.7.1
pytesseract==0.3.7
pytz==2022.1
requests-oauthlib==1.3.0
rsa==4.7.2
runcython3==0.2.7
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tomli==2.0.1
typing_extensions==4.2.0
virtualenv==20.4.7
wrapt==1.12.1
youtube-dl==2021.6.6
